--- features/principledetail/src/main/java/com/depromeet/team5/features/principledetail/event/Event.kt
@@ -1,13 +1,0 @@
-package com.depromeet.team5.features.principledetail.event
-
-
-sealed class PrincipleDetailEvent {
-
-    object Finish : PrincipleDetailEvent()
-
-    data class ShowErrorToast(val throwable: Throwable) : PrincipleDetailEvent()
-
-    object FinishAndShowToast : PrincipleDetailEvent()
-
-    object None : PrincipleDetailEvent()
-}
