+++ features/principlegroupdetail/src/main/java/com/depromeet/team5/features/principlegroupdetail/navigation/Navigation.kt
@@ -1,0 +1,45 @@
+package com.depromeet.team5.features.principlegroupdetail.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavOptions
+import androidx.navigation.compose.composable
+import androidx.navigation.toRoute
+import com.depromeet.team5.core.navigation.PrincipleType
+import com.depromeet.team5.features.principlegroupdetail.screen.PrincipleDetailRoute
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+data class PrincipleGroupDetail(
+    val groupId: Int,
+    val principleType: PrincipleType
+)
+
+fun NavController.navigateToPrincipleGroupDetail(
+    groupId: Int,
+    principleType: PrincipleType,
+    options: NavOptions? = null
+) {
+    navigate(
+        route = PrincipleGroupDetail(groupId, principleType),
+        navOptions = options
+    )
+}
+
+fun NavGraphBuilder.principleGroupDetail(
+    onBackPressed: () -> Unit,
+    onShowErrorToast: (Throwable) -> Unit,
+    onShowToast: (String) -> Unit
+) {
+    composable<PrincipleGroupDetail> { backstackEntry ->
+        val args = backstackEntry.toRoute<PrincipleGroupDetail>()
+
+        PrincipleDetailRoute(
+            principleType = args.principleType,
+            onBackPressed = onBackPressed,
+            onShowErrorToast = onShowErrorToast,
+            onShowToast = onShowToast
+        )
+    }
+}
