+++ features/principlegroupdetail/src/main/res/values/strings.xml
