+++ settings.gradle.kts
@@ -44,4 +44,4 @@ include(":features:feedback")
 include(":features:reasons")
 include(":core:logger")
 include(":core:ui")
-include(":features:principledetail")
+include(":features:principlegroupdetail")
